SWR fetching for listProjects from blocksync
