+++ package.json
@@ -79,6 +79,7 @@
     "reselect": "^4.0.0",
     "serve": "^11.3.2",
     "styled-components": "^5.1.1",
+    "swr": "^1.3.0",
     "typescript": "^3.9.7"
   },
   "devDependencies": {
