+++ src/modules/Entities/EntitiesExplorer/EntitiesExplorer.actions.test.ts
@@ -1,6 +1,6 @@
 import moment from 'moment'
 import * as SUT from './EntitiesExplorer.actions'
-import { EntityType, TermsOfUseType } from '../types'
+import { EntityType } from '../types'
 import { EntitiesExplorerActions } from './types'
 import mockStore from 'common/redux/mockStore'
 
@@ -25,326 +25,6 @@ beforeEach(() => {
 })
 
 describe('Entities Actions', () => {
-  describe('getEntities', () => {
-    // TOO enable when live api enabled again
-    it('should retrieve the entities from the api', async () => {
-      // when we call the action creator
-      await store.dispatch(SUT.getEntities())
-      const actions = store.getActions()
-
-      expect(actions.length).toEqual(2)
-
-      // then ... it should dispatch the correct actions
-      expect(actions.length).toEqual(2)
-
-      expect(actions[0].type).toEqual(
-        EntitiesExplorerActions.GetEntitiesPending,
-      )
-      expect(actions[1].type).toEqual(
-        EntitiesExplorerActions.GetEntitiesSuccess,
-      )
-      expect(actions[1].payload).toEqual([
-        {
-          name: 'Some Title',
-          description: 'Some Short Description',
-          type: EntityType.Project,
-          did: 'did:ixo:GfDZQaXJ9o2UKm4tGY2Wkh',
-          disputedClaimsCount: 0,
-          creatorDid: 'did:sov:EA1fV7PTbWG3aveDJZpgSn',
-          dateCreated: moment('2020-09-12T19:49:45Z'),
-          creatorName: 'Creator Display Name',
-          creatorLogo:
-            'https://pds_pandora.ixo.world/public/8520qk1ckqvkelkjfeg',
-          status: 'CREATED',
-          location: 'AR',
-          goal: undefined,
-          serviceProvidersCount: 10,
-          evaluatorsCount: 10,
-          requiredClaimsCount: 0,
-          successfulClaimsCount: 0,
-          pendingClaimsCount: 0,
-          rejectedClaimsCount: 0,
-          sdgs: ['5', '7'],
-          agentDids: [
-            'did:ixo:CB1idAyvNUsSEktkT3a5LY',
-            'did:ixo:NT1idAyvNUsSEktkT3a5LY',
-          ],
-          image: 'https://pds_pandora.ixo.world/public/sbujb0xg0dgkeljwtnc',
-          logo: 'https://pds_pandora.ixo.world/public/v7kvrycap9kf2ofnof',
-          ddoTags: [
-            {
-              name: 'Project Type',
-              tags: [
-                'Index',
-                'Accreditation',
-                'Accountability',
-                'Insurance Bond',
-              ],
-            },
-            {
-              name: 'SDG',
-              tags: [
-                'SDG3 – Good Health and Well-being',
-                'SDG15 – Life on Land',
-                'SDG16 – Peace, Justice and Strong Institutions',
-                'SDG17 – Partnerships for Goals',
-              ],
-            },
-            { name: 'Stage', tags: ['Planning'] },
-          ],
-          termsType: TermsOfUseType.OnceOffFee,
-          badges: ['https://somebadge.com', 'https://anotherbadge.com'],
-          version: '1.0.5',
-          claims: [
-            {
-              date: new Date('2020-11-26T19:24:18.167Z'),
-              location: {
-                long: 'someLong',
-                lat: 'someLat',
-              },
-              claimId: '1',
-              status: '0',
-              saDid: 'someSADid',
-              eaDid: 'someEADid',
-            },
-            {
-              date: new Date('2020-11-26T19:24:18.167Z'),
-              location: {
-                long: 'someLong',
-                lat: 'someLat',
-              },
-              claimId: '2',
-              status: '0',
-              saDid: 'someSADid',
-              eaDid: 'someEADid',
-            },
-            {
-              date: new Date('2020-11-26T19:24:18.167Z'),
-              location: {
-                long: 'someLong',
-                lat: 'someLat',
-              },
-              claimId: '3',
-              status: '0',
-              saDid: 'someSADid',
-              eaDid: 'someEADid',
-            },
-            {
-              date: new Date('2020-11-26T19:24:18.167Z'),
-              location: {
-                long: 'someLong',
-                lat: 'someLat',
-              },
-              claimId: '4',
-              status: '1',
-              saDid: 'someSADid',
-              eaDid: 'someEADid',
-            },
-          ],
-          liquidity: {
-            '@context': 'https://schema.ixo.world/liquidity/ipfs3r08webu2eou',
-            items: [
-              {
-                '@type': 'Alphabond',
-                id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdzzzz',
-              },
-            ],
-          },
-          entityClaims: {
-            '@context': 'https://schema.ixo.world/claims:3r08webu2eou',
-            items: [
-              {
-                '@id': 'template:did:2',
-                agents: [
-                  {
-                    autoApprove: true,
-                    credential: 'Credential 1',
-                    role: 'PO',
-                  },
-                  {
-                    autoApprove: false,
-                    credential: 'Credential 2',
-                    role: 'SA',
-                  },
-                  {
-                    autoApprove: true,
-                    credential: 'Credential 3',
-                    role: 'IA',
-                  },
-                ],
-                claimApproval: [
-                  {
-                    // '@type': 'Alphabond',
-                    // id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdzzzz',
-                    '@context': 'somecontext1',
-                    '@id': 'somelink1',
-                    criteria: [
-                      {
-                        attribute: 'criteria1',
-                        condition: 'condition1',
-                      },
-                      {
-                        attribute: 'criteria2',
-                        condition: 'condition2',
-                      },
-                      {
-                        attribute: 'criteria3',
-                        condition: 'condition3',
-                      },
-                      {
-                        attribute: 'criteria4',
-                        condition: 'condition4',
-                      },
-                    ],
-                  },
-                ],
-                claimEnrichment: [
-                  {
-                    '@context': 'somecontext1',
-                    '@id': 'somelink1',
-                    resources: [
-                      {
-                        productId: 'productId1',
-                        resource: 'res1',
-                      },
-                      {
-                        productId: 'productId2',
-                        resource: 'res2',
-                      },
-                      {
-                        productId: 'productId3',
-                        resource: 'res3',
-                      },
-                    ],
-                  },
-                  {
-                    '@context': 'somecontext2',
-                    '@id': 'somelink2',
-                    resources: [
-                      {
-                        productId: 'productId1',
-                        resource: 'res1',
-                      },
-                      {
-                        productId: 'productId2',
-                        resource: 'res2',
-                      },
-                    ],
-                  },
-                ],
-                claimEvaluation: [
-                  {
-                    '@context': 'somecontext1',
-                    '@id': 'somelinktocontext1',
-                    attributes: ['attr1', 'attr2'],
-                    methodology: 'somemethodology1',
-                  },
-                  {
-                    '@context': 'somecontext2',
-                    '@id': 'somelinktocontext2',
-                    attributes: ['attr1', 'attr2', 'attr3'],
-                    methodology: 'somemethodology2',
-                  },
-                ],
-                description: 'Some Claim Description 1',
-                endDate: '2020-10-22T00:00:00.000Z',
-                goal: 'Some Goal',
-                startDate: '2020-09-25T00:00:00.000Z',
-                targetMax: 45,
-                targetMin: 23,
-                title: 'Some Claim Title 1',
-                visibility: 'Private',
-              },
-            ],
-          },
-          linkedEntities: [
-            {
-              '@type': 'Investment',
-              id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdqq',
-            },
-            {
-              '@type': 'Oracle',
-              id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdtt',
-            },
-          ],
-        },
-        {
-          name: 'Some Claim Title',
-          description: 'Some Claim Description',
-          type: EntityType.Template,
-          did: 'did:ixo:BuXZQaXJ9o2UKm4tGYKLW',
-          disputedClaimsCount: 0,
-          creatorDid: 'did:sov:AB1fV7PTbWG3aveDJZpgSb',
-          dateCreated: moment('2020-08-12T19:49:45Z'),
-          creatorName: 'Creator Display Name',
-          creatorLogo:
-            'https://pds_pandora.ixo.world/public/8520qk1ckqvkelkjfeg',
-          status: 'COMPLETED',
-          serviceProvidersCount: 10,
-          evaluatorsCount: 10,
-          agentDids: [
-            'did:ixo:TB1idAyvNUsSEktkT3a5LY',
-            'did:ixo:VT1idAyvNUsSEktkT3a5LY',
-          ],
-          ddoTags: [
-            {
-              name: 'Project Type',
-              tags: [
-                'Index',
-                'Accreditation',
-                'Accountability',
-                'Insurance Bond',
-              ],
-            },
-            {
-              name: 'SDG',
-              tags: [
-                'SDG3 – Good Health and Well-being',
-                'SDG15 – Life on Land',
-                'SDG16 – Peace, Justice and Strong Institutions',
-                'SDG17 – Partnerships for Goals',
-              ],
-            },
-            { name: 'Stage', tags: ['Planning'] },
-          ],
-          termsType: TermsOfUseType.OnceOffFee,
-          badges: ['https://somebadge.com', 'https://anotherbadge.com'],
-          version: '1.0.5',
-          claims: [],
-          liquidity: {
-            '@context': 'https://schema.ixo.world/liquidity/ipfs3r08webu2eou',
-            items: [
-              {
-                '@type': 'Alphabond',
-                id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdzzzz',
-              },
-            ],
-          },
-          successfulClaimsCount: 0,
-          pendingClaimsCount: 0,
-          rejectedClaimsCount: 0,
-          requiredClaimsCount: 0,
-          sdgs: undefined,
-          location: undefined,
-          logo: undefined,
-          linkedEntities: [
-            {
-              '@type': 'Investment',
-              id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdqq',
-            },
-            {
-              '@type': 'Oracle',
-              id: 'did:sov:CYCc2xaJKrp8Yt947Nc6jdtt',
-            },
-          ],
-          entityClaims: undefined,
-          goal: undefined,
-          image: undefined,
-        },
-      ])
-    })
-  })
-
   describe('changeEntityType', () => {
     it('should create an action to set the entityType', () => {
       // when ... we call the changeEntityType action creator
