+++ src/modules/Entities/SelectedEntity/EntityExchange/Airdrop/Airdrop.container.tsx
@@ -1,20 +1,13 @@
 import React, { useEffect, useState } from 'react'
 import { RootState } from 'common/redux/types'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector } from 'react-redux'
 import DataCard from 'modules/Entities/EntitiesExplorer/components/EntityCard/AirdropCard/AirdropCard'
 import { EntityType, TermsOfUseType } from 'modules/Entities/types'
 import { ExplorerEntity } from 'modules/Entities/EntitiesExplorer/types'
-import { getEntities } from 'modules/Entities/EntitiesExplorer/EntitiesExplorer.actions'
 
 const Airdrop: React.FunctionComponent = () => {
-  const dispatch = useDispatch()
   const { entities } = useSelector((state: RootState) => state.entities)
   const [airdropList, setAirdropList] = useState<ExplorerEntity[]>([])
-
-  useEffect(() => {
-    dispatch(getEntities())
-    // eslint-disable-next-line
-  }, [])
 
   useEffect(() => {
     //  temporary placeholder
@@ -22,7 +15,7 @@ const Airdrop: React.FunctionComponent = () => {
     if (!entities) {
       return
     }
-    let filtered = entities
+    const filtered = entities
       .filter(entity => entity.type === EntityType.Project)
       .filter(entity =>
         entity.ddoTags.some(
