+++ src/services/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import EntitiesService from './entities.service'
+
+const Services: React.FC = (): JSX.Element => (
+  <>
+    <EntitiesService />
+  </>
+)
+
+export default Services
